requirements.txt
numpy==2.0.2
opencv-contrib-python==4.10.0.84
opencv-python==4.10.0.84
opencv-python-headless==4.10.0.84
pillow==10.4.0
